# feat: upgrade to npm3 and use npminstall