--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cnpm",
-  "version": "3.4.1",
+  "version": "4.0.0",
   "description": "cnpm: npm client for cnpmjs.org",
   "main": "bin/cnpm",
   "scripts": {