--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cnpm",
-  "version": "4.2.0-beta.3",
+  "version": "4.2.0-beta.4",
   "description": "cnpm: npm client for cnpmjs.org",
   "main": "bin/cnpm",
   "scripts": {
@@ -30,7 +30,7 @@
     "open": "~0.0.5",
     "node-gyp": "^3.3.0",
     "urllib": "^2.8.0",
-    "npminstall": "^1.2.5"
+    "npminstall": "^1.3.1"
   },
   "devDependencies": {
     "autod": "*",