--- a/parse_argv.js
+++ b/parse_argv.js
@@ -1,5 +1,5 @@
-/**!
- * Copyright(c) cnpmjs.org and other contributors.
+/**
+ * Copyright(c) cnpm and other contributors.
  * MIT Licensed
  *
  * Authors: