cnpm install 4x faster: `$ cnpm install cnpmjs.org`

- Before: `cnpm@3.4.1`

```bash
$ cnpm cache clean
$ time cnpm install cnpmjs.org

real	0m24.370s
user	0m12.982s
sys	0m4.170s
```

- After: `cnpm@4.2.0`

```bas
$ cnpm cache clean
$ time cnpm install cnpmjs.org --no-cache

[cnpmjs.org@*] installed at node_modules/.npminstall/cnpmjs.org/2.8.0/cnpmjs.org (0 packages, use 5s, speed 1.04MB/s, json 1.44MB, tarball 3.74MB)
All packages installed (215 packages installed from npm registry, use 5s, speed 1.03MB/s, json 266(1.44MB), tarball 3.74MB)

real	0m6.220s
user	0m4.871s
sys	0m2.072s
```

Why we can improve the install speed? Check out [npminstall](https://github.com/cnpm/npminstall) and take a look.