@fengmk2 2016-03-04
@dead-horse after the merge, we will publish v4 to latest version