@fengmk2 2016-03-05
Wait for shrinkwrap https://github.com/cnpm/npminstall/issues/37