@alsotang 2016-03-07
https://npm.taobao.org/mirrors

这个网址的东西以后还会维护吗？我看代码里面把它删了