@dead-horse 2016-03-07
@alsotang 在 https://github.com/cnpm/mirrors