@alsotang 2016-03-07
![image](https://cloud.githubusercontent.com/assets/1147375/13560653/a60a5252-e45c-11e5-9884-3432d2e9b8e7.png)

我弄错了。。这个 `mirrorsUrl` 在代码里是没有出现过的。