@dead-horse 2016-03-07
这个 mirrors 的环境变量 npminstall 会设置，所以这边都去掉了