@XGHeaven 2016-03-08
能否选择使用 `npm` 或者 `npminstall` 进行安装？

因为有很多我以前的项目是按照 `npm3` 的格式组织的，所以需要 `npm` 才可以。

还有希望添加进度条~~~