@fengmk2 2016-03-08
@XGHeaven 代码里面不是直接用 require 的吗？遵循 node_modules 约定就可以了。