@XGHeaven 2016-03-08
@fengmk2 话是这么说，可是这里会产生一个兼容问题。比如说有一个库通过 npm3 安装，那么下面的所有库都会 flat 的方式安装到这个目录。此时再用 npminstall 安装，会出现重新安装了这些依赖。这样就会产生冗余。有一些框架比如 node-weixin，就会产生问题。