@fengmk2 2016-03-08
node-weixin 看了代码 https://github.com/feeloc/node-weixin 也不会有问题啊