@XGHeaven 2016-03-08
@fengmk2 删除重新安装当然是可以了。可能觉得在这个时候用回原先的 npm 最为方便实在。实在不行，我建议可以再加一个 `cnpm-old` 用来支持原生 `npm`。

我说的是 `node-weixin` 是一个 organization，作者是田老大，他记得他说里面的 `node-weixin-config` 准备只兼容 npm3 的 flat 模式，也就是单例引入了。因为会产生冗余，所以有可能内部模块引入的 `node-weixin-config` 可能和我自己引入的不是一个实例，从而产生错误。