@fengmk2 2016-03-08
@XGHeaven 我跑 node-weixin-config 项目测试是 pass 的

![image](https://cloud.githubusercontent.com/assets/156269/13606364/ab781f6e-e586-11e5-9aec-bb306d480da6.png)