@alsotang 2016-03-08
为什么 node_modules 不能删掉重装？

2016-03-08 23:37 GMT+08:00 fengmk2 <notifications@github.com>:

> cc @JacksonTian <https://github.com/JacksonTian> 你会写出这么搓的模块？
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/cnpm/cnpm/pull/72#issuecomment-193827210>.
>