@dead-horse 2016-03-08
npminstall 安装好之后也是单例模式，但是依赖这个单例模式迟早踩坑