@fengmk2 2016-03-08
npm3 也不能保证单例的。