@XGHeaven 2016-03-08
npminstall确实是单例的。但是npm和npminstall混合使用可能就不保证了😲