@fengmk2 2016-03-08
sqlite3 安装也不再痛苦了 https://github.com/cnpm/npminstall/pull/47/files#diff-7541d61d4006fbd2288698bca7e8e83dR38