@XGHeaven 2016-03-08
@dead-horse 那还需要把以前的 node-modules 删掉。好麻烦的说 :joy:

还有能加一个进度条显示么？@fengmk2