@fengmk2 2016-03-09
不会加的，因为足够快，没必要进度条。如果你还是想继续用原来的，你直接安装cnpm3即可。

Sent from my iPhone

> On Mar 9, 2016, at 7:49 AM, Xu Tianyang <notifications@github.com> wrote:
> 
> @dead-horse 那还需要把以前的 node-modules 删掉。好麻烦的说 
> 
> 还有能加一个进度条显示么？@fengmk2
> 
> —
> Reply to this email directly or view it on GitHub.
>